#1 - Improve edge detection, use shorter wait intervals.
Simplify code, keep IO-agnostic methods in the separate file.
